#1522 A lot of warning when compile indigo-core-unit-tests
Fix warnings in render2d project.
Enable '-WX' (Treats all compiler warnings as errors.) for MSVC in render2d project

+++ core/indigo-core/molecule/molecule_dearom.h
@@ -76,7 +76,7 @@ namespace indigo
         {
             return _dearomParams;
         }
-        void setDearomatizationParams(int params)
+        void setDearomatizationParams(byte params)
         {
             _dearomParams = params;
         }
